backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Media
opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0

# AI APIs (OpenAI for Whisper + GPT-4V + text; can add anthropic later)
openai==1.55.3

# Utilities
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Background jobs (optional Redis)
arq==0.26.1
redis==5.2.1
